mqtt/requirements.txt
pymongo==3.6.1
paho-mqtt==1.5.1